Working version of the first semester

+++ lib/STL_Asm/STL_fixup_struct.h
@@ -1,10 +1,12 @@
 #ifndef STL_fixup_struct_
 #define STL_fixup_struct_
 
+#include <stdio.h>
+
 struct Fixup
 {
-    int address;
-    int labelIndex;
+    size_t address;
+    size_t labelIndex;
 };
 
 #endif /* STL_fixup_struct_ */
