+++ lib/STL_Asm/STL_label.h
@@ -3,17 +3,17 @@
 
 #include "../STL_Spu/include/STL_spu_struct_const.h"
 #include "STACK_FILE_TYPE_label.h"
-#include "../STL_stack/generic.h"
+#include "../STL_Stack/generic.h"
 #include "STACK_FILE_TYPE_fixup.h"
-#include "../STL_stack/generic.h"
+#include "../STL_Stack/generic.h"
 
 #include <stdio.h>
 #include <string.h>
 #include <assert.h>
 
-int SetLabel    (const char* string, const int ip, Stack_Label* labels);
+int SetLabel    (const char* string, const size_t ip, Stack_Label* labels);
 
-int SetLabelArg (const char* string, char* str, int* ip, Stack_Label* labels, Stack_Fixup* fixups);
+int SetLabelArg (const char* string, char* str, size_t* ip, Stack_Label* labels, Stack_Fixup* fixups);
 
 int DoFixups    (char* str, const Stack_Label* labels, const Stack_Fixup* fixups);
